Appease linter
An otherwise effective lint happens to be wrong in this specific case. Clippy is unhappy that I am dividing a length by the size of an element in an unsafe operation.

This is normally a footgun, however in this case, I am assembling a new slice from raw parts to reinterpret a byte slice as a slice of some other numeric value. This is perfectly legitimate, as the underlying bytes here represent elements of that numeric type, and thus this division is necessary to establish the correct length for the new slice.